realtime/python/requirements.txt
aiohttp>=3.9
httpx>=0.27
python-dotenv>=1.0
